Solve 12.b
Phew, this one was extremely tough for me. I'm not particular fond of my
solution, but incorporating a nice way of checking those diagonals sides
was very difficult. The special cases array is used to validating if
patterns of

```
X Y
Y X
```

has one or two corners in the y and x location. They have 2 there if and
only if the Y's share region.