Solve 13.a
I did not really read anything about Scala before jumping into it. And
my solution is definently not idiomatic as Scala as far as I can tell
promotes functional programming (no continue and break in standard for
loop), so my solve function would probably be a lot prettier in a
recursive format.

The solution is very very crude, and is easily optimizable, which I am
guessing will be needed for part 2.